:art: Make main file minimal, move everything to lib.rs
